Change traId to int

+++ Src/DfT.DTRO.ApiTests/ApiTests/Scenarios/Schema_3.3.0/3.3.0_UpdateDtroScenarios/3.3.0_HappyScenarios.cs
@@ -41,7 +41,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_3_0.DtroUpdateScenarios
             // Prepare DTRO
             string dtroCreationJson = fileName
                                     .GetJsonFromFile(schemaVersionToTest)
-                                    .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId);
+                                    .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId);
 
             // Send DTRO
             HttpResponseMessage dtroCreationResponse = await dtroCreationJson.SendJsonInDtroCreationRequestAsync(publisher);
@@ -87,7 +87,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_3_0.DtroUpdateScenarios
             // Prepare DTRO
             string dtroCreationJson = fileName
                                     .GetJsonFromFile(schemaVersionToTest)
-                                    .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId);
+                                    .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId);
 
             string tempFilePathForDtroCreation = dtroCreationJson.CreateDtroTempFile(fileName, publisher);
